211006 jaehyun
chore
회원가입 시 BMI계산식 추가
1. Member.java 신체정보 자료형 double로 변경
2. MemberController 수식추가

MEMBER_INFO.sql insert 구문 추가
join.jsp 선호운동 checkbox값 진행중(미완성)

+++ Sports/src/main/java/com/project/sports/domain/Member.java
@@ -17,11 +17,11 @@ public class Member {
 	private String HOMEADDRESS;
 	private String USER_EMAIL;
 	private String USER_EMAILDOMAIN;
-	private int USER_HEIGHT;	 //키
-	private int USER_PWEIGHT;	 //현재몸무게
-	private int USER_WWEIGHT;	 //목표몸무게
-	private int USER_BMI;		 //BMI(현재몸무게/(키*키)*10000)
-	private int USER_RMR;		 //기초대사량
+	private double USER_HEIGHT;	 //키
+	private double USER_PWEIGHT;	 //현재몸무게
+	private double USER_WWEIGHT;	 //목표몸무게
+	private double USER_BMI;		 //BMI(현재몸무게/(키*키)*10000)
+	private double USER_RMR;		 //기초대사량
 	//(남자:66.47+(13.75*현재몸무게)+(5*키)-(6.75*나이)
 	//(여자:655.1+(9.56*현재몸무게)+(1.85*키)-(4.68*나이)
 	private int USER_WATER;		 //물섭취량
@@ -125,34 +125,34 @@ public class Member {
 	public void setUSER_EMAILDOMAIN(String uSER_EMAILDOMAIN) {
 		USER_EMAILDOMAIN = uSER_EMAILDOMAIN;
 	}
-	public int getUSER_HEIGHT() {
+	public double getUSER_HEIGHT() {
 		return USER_HEIGHT;
 	}
-	public void setUSER_HEIGHT(int uSER_HEIGHT) {
+	public void setUSER_HEIGHT(double uSER_HEIGHT) {
 		USER_HEIGHT = uSER_HEIGHT;
 	}
-	public int getUSER_PWEIGHT() {
+	public double getUSER_PWEIGHT() {
 		return USER_PWEIGHT;
 	}
-	public void setUSER_PWEIGHT(int uSER_PWEIGHT) {
+	public void setUSER_PWEIGHT(double uSER_PWEIGHT) {
 		USER_PWEIGHT = uSER_PWEIGHT;
 	}
-	public int getUSER_WWEIGHT() {
+	public double getUSER_WWEIGHT() {
 		return USER_WWEIGHT;
 	}
-	public void setUSER_WWEIGHT(int uSER_WWEIGHT) {
+	public void setUSER_WWEIGHT(double uSER_WWEIGHT) {
 		USER_WWEIGHT = uSER_WWEIGHT;
 	}
-	public int getUSER_BMI() {
+	public double getUSER_BMI() {
 		return USER_BMI;
 	}
-	public void setUSER_BMI(int uSER_BMI) {
+	public void setUSER_BMI(double uSER_BMI) {
 		USER_BMI = uSER_BMI;
 	}
-	public int getUSER_RMR() {
+	public double getUSER_RMR() {
 		return USER_RMR;
 	}
-	public void setUSER_RMR(int uSER_RMR) {
+	public void setUSER_RMR(double uSER_RMR) {
 		USER_RMR = uSER_RMR;
 	}
 	public int getUSER_WATER() {
